fastapi-backend/requirements.txt
pydantic[email]
annotated-types==0.7.0
anyio==4.8.0
arrow==1.3.0
bcrypt==4.2.1
binaryornot==0.4.4
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cookiecutter==2.6.0
dnspython==2.7.0
fastapi==0.115.8
greenlet==3.1.1
h11==0.14.0
idna==3.10
Jinja2==3.1.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
motor==3.7.0
passlib==1.7.4
psycopg2==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
text-unidecode==1.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
APScheduler==3.10.1
requests==2.32.3
